+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2024-05-17-1
+
+Changes:
+
+- Improved imports/exports
+
 ## v2024-05-16-1
 
 Changes: